+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add `render_in_html` method.
+- Add code documentation.
+
 ## 0.1.0 (2025-06-28)
 
 - Add `render_to_string` method.